# fix(core): fix performance impact in mutation due to vector