Added a new benchmark for it. 
Before:
```
badger 2024/10/09 22:48:46 INFO: All 0 tables opened in 0s
badger 2024/10/09 22:48:46 INFO: Discard stats nextEmptySlot: 0
badger 2024/10/09 22:48:46 INFO: Set nextTxnTs to 0
<nil>
goos: linux
goarch: amd64
pkg: github.com/dgraph-io/dgraph/v24/worker
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
BenchmarkAddMutationWithIndex-8         badger 2024/10/09 22:48:46 INFO: All 0 tables opened in 0s
badger 2024/10/09 22:48:46 INFO: Discard stats nextEmptySlot: 0
badger 2024/10/09 22:48:46 INFO: Set nextTxnTs to 0
<nil>
badger 2024/10/09 22:48:46 INFO: All 0 tables opened in 0s
badger 2024/10/09 22:48:46 INFO: Discard stats nextEmptySlot: 0
badger 2024/10/09 22:48:46 INFO: Set nextTxnTs to 0
<nil>
badger 2024/10/09 22:48:47 INFO: All 0 tables opened in 0s
badger 2024/10/09 22:48:47 INFO: Discard stats nextEmptySlot: 0
badger 2024/10/09 22:48:47 INFO: Set nextTxnTs to 0
<nil>
badger 2024/10/09 22:48:48 INFO: All 0 tables opened in 0s
badger 2024/10/09 22:48:48 INFO: Discard stats nextEmptySlot: 0
badger 2024/10/09 22:48:48 INFO: Set nextTxnTs to 0
<nil>
   64393             17490 ns/op
PASS
ok      github.com/dgraph-io/dgraph/v24/worker  2.467s
```
After:
```
 badger 2024/10/09 22:48:30 INFO: All 0 tables opened in 0s                                                                                                                                                                                                                                                                                                                                      
badger 2024/10/09 22:48:30 INFO: Discard stats nextEmptySlot: 0                                                                                                                                                                                                                                                                                                                                 
badger 2024/10/09 22:48:30 INFO: Set nextTxnTs to 0                                                                                                                                                                                                                                                                                                                                             
<nil>                                                                                                                                                                                                                                                                                                                                                                                           
goos: linux                                                                                                                                                                                                                                                                                                                                                                                     
goarch: amd64                                                                                                                                                                                                                                                                                                                                                                                   
pkg: github.com/dgraph-io/dgraph/v24/worker                                                                                                                                                                                                                                                                                                                                                     
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz                                                                                                                                                                                                                                                                                                                                             
BenchmarkAddMutationWithIndex-8         badger 2024/10/09 22:48:30 INFO: All 0 tables opened in 0s                                                                                                                                                                                                                                                                                              
badger 2024/10/09 22:48:30 INFO: Discard stats nextEmptySlot: 0                                                                                                                                                                                                                                                                                                                                 
badger 2024/10/09 22:48:30 INFO: Set nextTxnTs to 0                                                                                                                                                                                                                                                                                                                                             
<nil>                                                                                                                                                                                                                                                                                                                                                                                           
badger 2024/10/09 22:48:30 INFO: All 0 tables opened in 0s                                                                                                                                                                                                                                                                                                                                      
badger 2024/10/09 22:48:30 INFO: Discard stats nextEmptySlot: 0                                                                                                                                                                                                                                                                                                                                 
badger 2024/10/09 22:48:30 INFO: Set nextTxnTs to 0                                                                                                                                                                                                                                                                                                                                             
<nil>                                                                                                                                                                                                                                                                                                                                                                                           
badger 2024/10/09 22:48:30 INFO: All 0 tables opened in 0s                                                                                                                                                                                                                                                                                                                                      
badger 2024/10/09 22:48:30 INFO: Discard stats nextEmptySlot: 0                                                                                                                                                                                                                                                                                                                                 
badger 2024/10/09 22:48:30 INFO: Set nextTxnTs to 0                                                                                                                                                                                                                                                                                                                                             
<nil>                                                                                                                                                                                                                                                                                                                                                                                           
  102348             11774 ns/op                                                                                                                                                                                                                                                                                                                                                                
PASS                                                                                                                                                                                                                                                                                                                                                                                            
ok      github.com/dgraph-io/dgraph/v24/worker  1.428s     
```